requirements.txt
requests>=2.31.0
psutil>=5.9.0
python-dotenv>=1.0.0
